Add wait_with_timeout methods on CQ.
These require mutable access to the SQ as well as the CQ, but they only hold
the SQ long enough to submit the timeout. Therefore, they manage to wait with
timeout while not locking the SQ down during waiting, and also only performing
one syscall.

There is a runtime check that the SQ corresponds to this CQ (by comparing the
address of their io_uring instances). These methods will panic if you try to
pass a CQ and SQ that don't correspond to the same io_uring instance.