Change RegisteredFd system
RegisteredFd now tracks both an index and the actual RawFd; it's only 8
bytes total so this is fine. That way, a user holding a RegisteredFd can
associate it to a real Fd if they need.

We no longer panic when trying to do IO on a placeholder FD; it's better
and more flexible for users to just get an IO error if they do that.